Move `core/applicationstate` into `store/application-state`
- add `src/store/application-state.js`
- deprecate notice for `src/app/core/applicationstate.js`
